server: split off compare from ldap_handler
