+++ server/src/infra/ldap/mod.rs
@@ -1,3 +1,4 @@
+pub mod compare;
 pub mod handler;
 pub mod password;
 pub mod search;
